compose/local/swiftsrv/requirements.txt
caterpillar@git+https://github.com/MatrixEditor/caterpillar.git
umbrella-py
